Changed int i to size_t

+++ 0x1D-search_algorithms/0-linear.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+
 #include "search_algos.h"
 
 /**
@@ -16,7 +17,7 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 	int index = -1;
 
 	if (!array)
